+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix issue [#347](https://github.com/intersystems/language-server/issues/347): Setting a variable's subscript should not affect the type of that variable
 - Fix issue [#348](https://github.com/intersystems/language-server/issues/348): Support intellisense for method arguments that don't have a declared type
 - Fix issue [#349](https://github.com/intersystems/language-server/issues/349): Automatically provide intellisense for `%sqlcontext` variable
+- Fix issue [#350](https://github.com/intersystems/language-server/issues/350): Use more granular symbols for class members in Outline
 
 ## [2.6.1] - 2024-08-30
 - Fix issue [#343](https://github.com/intersystems/language-server/issues/343): foldingRange fails on some C-style block comments