+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2.6.2] - 2024-XX-XX
+## [2.6.2] - 2024-10-07
 - Fix issue [#345](https://github.com/intersystems/language-server/issues/345): Add intellisense for variables passed by reference as method arguments
 - Fix issue [#347](https://github.com/intersystems/language-server/issues/347): Setting a variable's subscript should not affect the type of that variable
 - Fix issue [#348](https://github.com/intersystems/language-server/issues/348): Support intellisense for method arguments that don't have a declared type